feat: physical_cards page now can connect serial cards
