+++ nuxt.config.ts
@@ -2,4 +2,12 @@
 export default defineNuxtConfig({
   modules: ["nuxt-auth-utils", "@nuxt/ui"],
   css: ["~/assets/css/main.css"],
+  app: {
+    head: {
+      title: "Ronimisliidu registri otsing",
+      htmlAttrs: {
+        lang: "et",
+      },
+    },
+  },
 });
