+++ server/api/check.get.ts
@@ -1,30 +1,12 @@
-export default defineEventHandler(async (event): Promise<CheckDto> => {
+export default defineEventHandler(async (event): Promise<ClimberDto> => {
   const { id } = getQuery(event);
 
   if (typeof id != "string" || !isIdCodeValid(id)) {
-    return {
-      success: false,
-      id,
-      message: "Invalid id code",
-    };
+    throw createError({
+      status: 400,
+      statusMessage: "Invalid id code",
+    });
   }
 
-  try {
-    const result = await findExamById(id);
-
-    return {
-      success: true,
-      ...result,
-    };
-  } catch (err) {
-    let message = undefined;
-    if (typeof err == "object" && err != null && "message" in err) {
-      message = err.message;
-    }
-    return {
-      id,
-      success: false,
-      message,
-    };
-  }
+  return await findExamById(id);
 });
